add CategoriesFilterComponent
