+++ src/components/categories-filter/categories-filter.module.css
@@ -1,0 +1,44 @@
+:global :root.dark {
+  --products-filter-category-button-text-color: var(--main-white-color);
+}
+
+:global :root.light {
+  --products-filter-category-button-text-color: var(--sixth-grey-color);
+}
+
+.categoryFilterWrapper {
+  display: flex;
+  gap: 10px;
+  justify-content: flex-end;
+}
+
+.categoryFilterCheckbox {
+  position: absolute;
+  appearance: none;
+}
+
+.categoryFilterButton {
+  line-height: 1.3;
+  font-weight: 600;
+  color: var(--products-filter-category-button-text-color);
+}
+
+.categoryFilterButton:hover {
+  background-color: var(--main-red-hover-color);
+  border-color: var(--main-red-hover-color);
+  color: var(--main-white-color);
+}
+
+.categoryFilterCheckbox:checked + .categoryFilterButton {
+  background-color: var(--main-red-color);
+  border-color: var(--main-red-color);
+  color: var(--main-white-color);
+}
+
+@media (max-width: 768px) {
+
+  .categoryFilterWrapper {
+    justify-content: center;
+    flex-wrap: wrap;
+  }
+}
