+++ src/helpers/categoriesFilterHelper.ts
